pyproject.toml
[tool.poetry]
name = "coursework-oop"
version = "0.1.0"
description = "coursework oop"
authors = ["aleksejbelkin1017 <aleksejbelkin1017@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv|.*/conftest.py'

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(\.git)
'''
